requirements.txt
streamlit~=1.45.1
python-docx~=1.1.2
openai~=1.88.0

